requirements.txt
python-telegram-bot==13.8.1
PyYAML
fastapi
requests
torch
dtrx
